feat: "Redis 연동을 위한 클래스 및 설정 추가"
application.yml - redis 연동 설정
RedisRepositoryConfig - redis 서버 연결 설정

Related to: #57, #75, #76
