+++ itty-backend/src/main/resources/application.yml
@@ -18,7 +18,11 @@ spring:
     redis:
       host: localhost
       port: 6379
+#      expiration_time: 2 * 24 * 60 * 1000L
+  cache:
+#    redis 캐시를 사용하기 위해 cache 타입을 redis로 설정
+    type: redis
 
 token:
-#  expiration_time: 86400000 #1일
+  expiration_time: 86400000 #1일
   secret: k/3f0rHITyFhkBQWfa7rlSAQ3s07wIgj7pUGgq5Q4EXbihvcPPapnHnKQAWV+u3mXJzIBQ9Cm++F0JPEHIv80g==
